# Allow stale data back in GetSpanInfo rule