@ndmitchell 2020-06-09
I think the problem is that `useWithStale` isn't really using stale results. It's using up to the minute results, but the last working up to the minute results, hence I always get tripped up by it. Definitely the right change to make.

But why is removing the early cutoff from the HiFile the right thing to do? Before if the interface hash was the same twice be cut off computation. That seems valuable in some cases, and I imagine a case your benchmark doesn't hit? Generally, having a cutoff hash or similar is very cheap, provided computing the hash is itself cheap.