@pepeiborra 2020-06-09
> But why is removing the early cutoff from the HiFile the right thing to do? Before if the interface hash was the same twice be cut off computation. That seems valuable in some cases, and I imagine a case your benchmark doesn't hit? Generally, having a cutoff hash or similar is very cheap, provided computing the hash is itself cheap.

Just a desire to minimize things. We already get cut off if the `GetModTime` is the same for the interface and all its dependencies, right? But I guess that doesn't cover comments and whitespace edits, and given that the hash is likely cheap in this case, I'm happy to drop it. Thanks!